feat: channel access token batch separation
