+++ src/application/services/channelAccessTokenService.ts
@@ -1,16 +1,22 @@
 import { Injectable, Logger } from '@nestjs/common';
-import * as jose from 'node-jose';
 import { fetchChannelAccessTokenResponseDto } from '../dto/channelAccessTokenDto';
 import { LineMessagingApiService } from 'src/infrastructure/api/line/lineMessagingApiService';
 import { IChannelAccessTokenService } from 'src/domain/interfaces/services/channelAccessTokenService';
+import { ChannelAccessTokenRepository } from 'src/infrastructure/persistence/repositories/channelAccessTokenRepository';
+import { generateJwt } from 'src/domain/useCase/jwt';
 
 // Log message constants
 const REQUEST_FETCH_CHANNEL_ACCESS_TOKEN_LOG =
   'Requesting fetch channel access token.';
-const REQUEST_FETCH_CHANNEL_ACCESS_TOKEN_FAILED_LOG =
-  'Failed to fetch channel access token.';
-const GENERATE_JWT_FAILED_LOG = 'Failed to generate JWT.';
 const GENERATE_JWT_LOG = 'Trying to generate JWT.';
+const GENERATE_JWT_FAILED_LOG = 'Failed to generate JWT.';
+const FETCH_CHANNEL_ACCESS_TOKEN_LOG = 'Trying to fetch channel access token.';
+const FETCH_CHANNEL_ACCESS_TOKEN_FAILED_LOG =
+  'Failed to fetch channel access token.';
+const UPDATE_CHANNEL_ACCESS_TOKEN_BATCH_SUCCESS_LOG =
+  'Successfully updated channel access token.';
+const UPDATE_CHANNEL_ACCESS_TOKEN_BATCH_FAILED_LOG =
+  'Failed to update channel access token.';
 
 /**
  * Channel access token service
@@ -21,58 +27,38 @@ export class ChannelAccessTokenService implements IChannelAccessTokenService {
 
   constructor(
     private readonly lineMessagingApiService: LineMessagingApiService,
+    private readonly channelAccessTokenRepository: ChannelAccessTokenRepository,
   ) {}
 
   /**
    * Fetch channel access token.
    * @returns channel access token
    */
-  async fetchChannelAccessToken(): Promise<fetchChannelAccessTokenResponseDto> {
+  async fetchChannelAccessToken(): Promise<void> {
     this.logger.log(REQUEST_FETCH_CHANNEL_ACCESS_TOKEN_LOG);
     try {
-      const jwt = await this.generateJwt();
-      return this.fetchChannelAccessTokenWithJwt(jwt);
+      let jwt: string;
+      try {
+        this.logger.log(GENERATE_JWT_LOG);
+        jwt = await generateJwt();
+      } catch (err) {
+        this.logger.error(GENERATE_JWT_FAILED_LOG, err.stack);
+        throw err;
+      }
+
+      const tokenResponse = await this.fetchChannelAccessTokenWithJwt(jwt);
+      this.logger.log(FETCH_CHANNEL_ACCESS_TOKEN_LOG);
+
+      await this.channelAccessTokenRepository.putChannelAccessToken(
+        tokenResponse.access_token,
+        tokenResponse.key_id,
+      );
+      this.logger.log(UPDATE_CHANNEL_ACCESS_TOKEN_BATCH_SUCCESS_LOG);
     } catch (err) {
       this.logger.error(
-        REQUEST_FETCH_CHANNEL_ACCESS_TOKEN_FAILED_LOG,
+        UPDATE_CHANNEL_ACCESS_TOKEN_BATCH_FAILED_LOG,
         err.stack,
       );
-      throw err;
-    }
-  }
-
-  /**
-   * Generate JWT.
-   * @returns JWT
-   */
-  private async generateJwt(): Promise<string> {
-    this.logger.log(GENERATE_JWT_LOG);
-
-    const privateKey = process.env.LINE_QUALE_QUICK_ALERT_SECRET_KEY;
-
-    const header = {
-      alg: 'RS256',
-      typ: 'JWT',
-      kid: process.env.LINE_QUALE_QUICK_ALERT_KID,
-    };
-
-    const payload = {
-      iss: process.env.LINE_QUALE_QUICK_ALERT_ISS,
-      sub: process.env.LINE_QUALE_QUICK_ALERT_SUB,
-      aud: 'https://api.line.me/',
-      exp: Math.floor(new Date().getTime() / 1000) + 60 * 30,
-      token_exp: 60 * 60 * 24 * 30,
-    };
-
-    try {
-      return await jose.JWS.createSign(
-        { format: 'compact', fields: header },
-        JSON.parse(privateKey),
-      )
-        .update(JSON.stringify(payload))
-        .final();
-    } catch (err) {
-      this.logger.error(GENERATE_JWT_FAILED_LOG, err.stack);
       throw err;
     }
   }
@@ -85,6 +71,11 @@ export class ChannelAccessTokenService implements IChannelAccessTokenService {
   private async fetchChannelAccessTokenWithJwt(
     jwt: string,
   ): Promise<fetchChannelAccessTokenResponseDto> {
-    return await this.lineMessagingApiService.fetchChannelAccessToken(jwt);
+    try {
+      return await this.lineMessagingApiService.fetchChannelAccessToken(jwt);
+    } catch (err) {
+      this.logger.error(FETCH_CHANNEL_ACCESS_TOKEN_FAILED_LOG, err.stack);
+      throw err;
+    }
   }
 }
