+++ src/domain/interfaces/services/channelAccessTokenService.ts
@@ -1,8 +1,6 @@
-import { fetchChannelAccessTokenResponseDto } from 'src/application/dto/channelAccessTokenDto';
-
 /**
  * Channel access token service interface
  */
 export interface IChannelAccessTokenService {
-  fetchChannelAccessToken(): Promise<fetchChannelAccessTokenResponseDto>;
+  fetchChannelAccessToken(): Promise<void>;
 }
